SP-934: Fix type in tests

+++ src/test/java/com/bitpay/sdk/model/invoice/MinerFeesItemTest.java
@@ -13,7 +13,7 @@ public class MinerFeesItemTest {
     @Test
     public void it_should_manipulate_satoshisPerByte() {
         // given
-        Integer expected = 765;
+        Double expected = 765;
         MinerFeesItem testedClass = this.getTestedClass();
 
         // when
